Move packSharedMemoryAlloc to LLVMGPU/Utils
