+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.h
@@ -1,5 +1,5 @@
 // Copyright 2023 The IREE Authors
-//
+//Utils/
 // Licensed under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
@@ -17,18 +17,6 @@
 
 namespace mlir {
 namespace iree_compiler {
-
-//===----------------------------------------------------------------------===//
-// Definitions and Utilities
-//===----------------------------------------------------------------------===//
-
-/// Look for allocs in shared memory space with overlapping liveness,
-/// group them, and then pack all the allocations in each group into one i8
-/// alloc.
-///
-/// Also adds barriers to make sure we are done writing/reading
-/// from the previous alias group before starting a new one.
-void packSharedMemoryAlloc(func::FuncOp funcOp);
 
 //===----------------------------------------------------------------------===//
 // Passes
